+++ android/src/main/kotlin/com/smileidentity/flutter/SmileIDSmartSelfieAuthentication.kt
@@ -11,6 +11,7 @@ import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.StandardMessageCodec
 import io.flutter.plugin.platform.PlatformView
 import io.flutter.plugin.platform.PlatformViewFactory
+import kotlinx.collections.immutable.toImmutableMap
 
 internal class SmileIDSmartSelfieAuthentication private constructor(
     context: Context,
@@ -24,11 +25,13 @@ internal class SmileIDSmartSelfieAuthentication private constructor(
 
     @Composable
     override fun Content(args: Map<String, Any?>) {
+        val partnerParams = args["partnerParams"] as? Map<String, String> ?: emptyMap()
         SmileID.SmartSelfieAuthentication(
             userId = args["userId"] as? String ?: randomUserId(),
             jobId = args["jobId"] as? String ?: randomJobId(),
             allowAgentMode = args["allowAgentMode"] as? Boolean ?: false,
             showAttribution = args["showAttribution"] as? Boolean ?: true,
+            partnerParams = partnerParams.toImmutableMap(),
         ) {
             when (it) {
                 is SmileIDResult.Success -> onSuccess(it.data)