+++ example/lib/main.dart
@@ -119,7 +119,6 @@ class MainContent extends StatelessWidget {
                 body: SmileIDDocumentVerification(
                   countryCode: "GH",
                   documentType: "DRIVERS_LICENSE",
-                  allowGalleryUpload: true,
                   onSuccess: (String? result) {
                     // Your success handling logic
                     final snackBar = SnackBar(content: Text("Success: $result"));