+++ ios/Classes/SmileIDDocumentVerification.swift
@@ -33,9 +33,11 @@ class SmileIDDocumentVerification : NSObject, FlutterPlatformView, DocumentVerif
             idAspectRatio: args["idAspectRatio"] as? Double,
             bypassSelfieCaptureWithFile: bypassSelfieCaptureWithFile,
             captureBothSides: args["captureBothSides"] as? Bool ?? true,
+            allowAgentMode: args["allowAgentMode"] as? Bool ?? false,
             allowGalleryUpload: args["allowGalleryUpload"] as? Bool ?? false,
             showInstructions: args["showInstructions"] as? Bool ?? true,
             showAttribution: args["showAttribution"] as? Bool ?? true,
+            partnerParams: args["partnerParams"] as? [String: String] ?? [:],
             delegate: self
         )
         let childViewController = UIHostingController(rootView: screen)