+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# 10.0.0-beta06
+- [Android] Added `extras` as optional params on all job types
+- [Android] Added `idAuthorityBypassPhoto` on Sandbox BiometricKYC jobs
+- [Android] Added `allowAgentMode` option on Document Verification and Enhanced Document Verification
+- [Flutter] Fixed wrong iOS decoding bug on success
+
 # 10.0.0-beta05
 - [Android] Fixed retry document submission on failed document submission
 - [Android] Fixed missing entered key in BiometricKYC